generalize to gigaviewer
