--- src/ja/ichicomi/src/eu/kanade/tachiyomi/extension/ja/ichicomi/Dto.kt
@@ -1,20 +1,0 @@
-package eu.kanade.tachiyomi.extension.ja.ichicomi
-
-import eu.kanade.tachiyomi.multisrc.gigaviewer.GigaViewerPage
-import kotlinx.serialization.Serializable
-
-@Serializable
-class PageListDto(
-    val readableProduct: ReadableProductDto,
-)
-
-@Serializable
-class ReadableProductDto(
-    val pageStructure: PageStructureDto,
-)
-
-@Serializable
-class PageStructureDto(
-    val pages: List<GigaViewerPage>,
-    val choJuGiga: String? = null,
-)
