+++ src/ja/kuragebunch/src/eu/kanade/tachiyomi/extension/ja/kuragebunch/KurageBunch.kt
@@ -2,16 +2,9 @@ package eu.kanade.tachiyomi.extension.ja.kuragebunch
 
 import eu.kanade.tachiyomi.multisrc.gigaviewer.GigaViewer
 import eu.kanade.tachiyomi.network.GET
-import eu.kanade.tachiyomi.source.model.Page
 import eu.kanade.tachiyomi.source.model.SManga
-import keiyoushi.utils.parseAs
-import kotlinx.serialization.SerializationException
-import okhttp3.HttpUrl.Companion.toHttpUrl
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import okhttp3.Request
-import okhttp3.Response
-import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
 
 class KurageBunch : GigaViewer(
@@ -41,41 +34,6 @@ class KurageBunch : GigaViewer(
     }
 
     override fun chapterListSelector(): String = "li.episode"
-
-    override fun pageListParse(document: Document): List<Page> {
-        val episodeJson = document.selectFirst("script#episode-json")?.attr("data-value")
-            ?: return emptyList()
-
-        val episode = try {
-            episodeJson.parseAs<PageListDto>()
-        } catch (e: SerializationException) {
-            throw Exception("このチャプターは非公開です\nChapter is not available!")
-        }
-        val isScrambled = episode.readableProduct.pageStructure.choJuGiga == "baku"
-
-        return episode.readableProduct.pageStructure.pages
-            .filter { it.type == "main" }
-            .mapIndexed { i, page ->
-                val imageUrl = page.src.toHttpUrl().newBuilder().apply {
-                    if (isScrambled) {
-                        addQueryParameter("width", page.width.toString())
-                        addQueryParameter("height", page.height.toString())
-                        addQueryParameter("descramble", "true")
-                    }
-                }.toString()
-                Page(i, document.location(), imageUrl)
-            }
-    }
-
-    override fun imageIntercept(chain: Interceptor.Chain): Response {
-        val request = chain.request()
-        val isScrambled = request.url.queryParameter("descramble") == "true"
-
-        if (!isScrambled) {
-            return chain.proceed(request)
-        }
-        return super.imageIntercept(chain)
-    }
 
     override fun getCollections(): List<Collection> = listOf(
         Collection("くらげバンチ", "kuragebunch"),
